run tests with submodule data, if data doesn't exist, print out skipp…

+++ src/main/java/org/example/Sum.java
@@ -6,7 +6,7 @@ package org.example;
  */
 public class Sum {
 
-    public long sumNumbers(int[] arr) {
+    public long sumNumbers(short[] arr) {
         long sumValue = 0;
         for (int i = 0; i < arr.length; i++) {
             sumValue += arr[i];
